+++ memloom/pipeline/__init__.py
@@ -1,6 +1,7 @@
 """Pipeline package: transform records after collection."""
 from .dedup import Deduper
+from .denoise import Denoiser, DenoiserStats
 from .privacy import PrivacyFilter, PrivacyStats
 from .tag import tag_record
 
-__all__ = ["PrivacyFilter", "PrivacyStats", "Deduper", "tag_record"]
+__all__ = ["PrivacyFilter", "PrivacyStats", "Deduper", "Denoiser", "DenoiserStats", "tag_record"]